Make navbar stay at the top of the screen. Refactor navbar darken-background logic
