+++ src/Shared/NavBar/NavBar.css
@@ -1,9 +1,15 @@
 .navbar {
     display: flex;
+    position: fixed;
+    height: var(--nav-bar-height);
+    width: 100%;
     background-color: var(--tertiary-color);
     align-items: center;
     font-size: 16px;
     padding: 1rem;
+    top: 0;
+    left: 0;
+    z-index: 2;
 
     .navbar-layout {
         display: flex;
@@ -61,4 +67,19 @@
             box-sizing: border-box;
         }
     }
+}
+
+.darken-background {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: calc(100vh);
+    background-color: rgba(0, 0, 0, 0);
+    z-index: 1;
+}
+
+.darken-background-active {
+    background-color: rgba(0, 0, 0, 0.5);
+    transition: background-color 0.3s;
 }
