+++ src/Shared/SearchForUser/SearchForUser.vue
@@ -26,17 +26,16 @@
             <p v-if="query.length === 0">Find a friend by searching for their name</p>
         </div>
     </div>
-    <div v-show="showSearchResults"
-        class="darken-background"
-        :class="{ 'darken-background-active': isDarkenActive }"></div>
 </template>
 
 <script setup lang="ts">
-import { nextTick, ref } from 'vue';
+import { ref } from 'vue';
 import { UserSearchClient } from '../Clients/UserSearchClient';
 import type { UserState } from '../Models/UserState';
 import { useRouter } from 'vue-router';
 import { useUserStateStore } from '../UserStateStore';
+
+const emit = defineEmits(['darken', 'lighten']);
 
 const router = useRouter();
 const userStateStore = useUserStateStore();
@@ -48,7 +47,6 @@ const debounceTimeout = ref();
 const searchResults = ref([] as UserState[]);
 const noSearchResults = ref(false);
 const showSearchResults = ref(false);
-const isDarkenActive = ref(false);
 // Dictionary for query -> response to store search results and prevent excessive API calls
 // Useful if the user makes a typo or queries the same thing multiple times
 const cache: Record<string, UserState[]> = {};
@@ -109,20 +107,13 @@ function close(event : KeyboardEvent) {
     target.blur();
 }
 
-async function handleFocusIn() {
+function handleFocusIn() {
     showSearchResults.value = true;
-
-    // Wait for DOM to render darken-background div
-    await nextTick();
-    
-    // Delay the activation of the darken transition to ensure div is ready
-    setTimeout(() => {
-        isDarkenActive.value = true;
-    }, 10);
+    emit('darken');
 }
 
 function handleFocusOut() {
     showSearchResults.value = false;
-    isDarkenActive.value = false;
+    emit('lighten');
 }
 </script>
