+++ src/assets/base.css
@@ -7,6 +7,7 @@
   --quaternary-color: #F5EFEB;
   --transparent-color: #567c8d4a;
   --transparent-hover-color: #567c8d78;
+  --nav-bar-height: 80px;
 }
 
 *,
